LayoutTests/http/tests/ssl/applepay/ApplePayPaymentMethodUpdateEvent.https.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="/js-test-resources/ui-helper.js"></script>
<script src="/resources/js-test-pre.js"></script>
<script src="/resources/payment-request.js"></script>
</head>
<body>
<script>

description("Test the applepaypaymentmethodchanged event.");

window.jsTestIsAsync = true;

function validPaymentMethod() {
    return {
        supportedMethods: 'https://apple.com/apple-pay',
        data: {
            version: 2,
            merchantIdentifier: '',
            countryCode: 'US',
            supportedNetworks: ['visa', 'masterCard'],
            merchantCapabilities: ['supports3DS'],
        },
    }
}

function validPaymentDetails() {
    return {
        total: { label: 'Your label', amount: { currency: 'USD', value: '10.00' } },
    }
}

async function go() {
    debug("Testing ApplePayMerchantValidationEvent");
    debug("");

    expectedPaymentMethod = {
        displayName: 'test',
        network: 'test',
        type: 'prepaid',
        paymentPass: {
            primaryAccountIdentifier: 'test',
            primaryAccountNumberSuffix: 'test',
            deviceAccountIdentifier: 'test',
            deviceAccountNumberSuffix: 'test',
            activationState: 'requiresActivation',
        },
    };

    var paymentRequest = new PaymentRequest([validPaymentMethod()], validPaymentDetails());
    paymentRequest.onapplepaypaymentmethodchanged = (event) => {
        shouldBe("event.paymentMethod.displayName", "expectedPaymentMethod.displayName");
        shouldBe("event.paymentMethod.network", "expectedPaymentMethod.network");
        shouldBe("event.paymentMethod.type", "expectedPaymentMethod.type");
        shouldBe("event.paymentMethod.paymentPass.primaryAccountIdentifier", "expectedPaymentMethod.paymentPass.primaryAccountIdentifier");
        shouldBe("event.paymentMethod.paymentPass.primaryAccountNumberSuffix", "expectedPaymentMethod.paymentPass.primaryAccountNumberSuffix");
        shouldBe("event.paymentMethod.paymentPass.deviceAccountIdentifier", "expectedPaymentMethod.paymentPass.deviceAccountIdentifier");
        shouldBe("event.paymentMethod.paymentPass.deviceAccountNumberSuffix", "expectedPaymentMethod.paymentPass.deviceAccountNumberSuffix");
        shouldBe("event.paymentMethod.paymentPass.activationState", "expectedPaymentMethod.paymentPass.activationState");

        document.querySelector("button").remove();
        finishJSTest();
    };

    activateThen(() => {
        paymentRequest.show();
        internals.mockPaymentCoordinator.changePaymentMethod(expectedPaymentMethod);
    });
}

window.onload = function() {
    var button = document.querySelector("button");
    let x = button.offsetLeft + 2;
    let y = button.offsetTop + 2;
    UIHelper.activateAt(x, y);
}

</script>
<button onclick='go()'>Go</button>
<script src="/resources/js-test-post.js"></script>
</body>
</html>
