LayoutTests/http/tests/ssl/applepay/ApplePayPaymentMethodUpdateEvent.https-expected.txt

Test the applepaypaymentmethodchanged event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ApplePayMerchantValidationEvent

PASS event.paymentMethod.displayName is expectedPaymentMethod.displayName
PASS event.paymentMethod.network is expectedPaymentMethod.network
PASS event.paymentMethod.type is expectedPaymentMethod.type
PASS event.paymentMethod.paymentPass.primaryAccountIdentifier is expectedPaymentMethod.paymentPass.primaryAccountIdentifier
PASS event.paymentMethod.paymentPass.primaryAccountNumberSuffix is expectedPaymentMethod.paymentPass.primaryAccountNumberSuffix
PASS event.paymentMethod.paymentPass.deviceAccountIdentifier is expectedPaymentMethod.paymentPass.deviceAccountIdentifier
PASS event.paymentMethod.paymentPass.deviceAccountNumberSuffix is expectedPaymentMethod.paymentPass.deviceAccountNumberSuffix
PASS event.paymentMethod.paymentPass.activationState is expectedPaymentMethod.paymentPass.activationState
PASS successfullyParsed is true

TEST COMPLETE
